move combination queries to repository instead of dataProvider
