+++ src/PrestaShopBundle/Resources/config/services/adapter/data_provider.yml
@@ -80,8 +80,6 @@ services:
         class: PrestaShop\PrestaShop\Adapter\CombinationDataProvider
         arguments:
             - "@prestashop.core.localization.locale.context_locale"
-            - '@doctrine.dbal.default_connection'
-            - '%database_prefix%'
 
     prestashop.adapter.data_provider.cms:
         class: PrestaShop\PrestaShop\Adapter\CMS\CMSDataProvider
